Woops

+++ src/layout.rs
@@ -380,8 +380,7 @@ pub fn thread3(rx: RChan, thread_number: usize, config: Config, layout_config: L
         for (mini_batch_size, rq) in rqs {
             // XXX actually clean the padded values of past so that subsequent
             // calls can get a chance to have a better padding (+ correct attention mask).
-
-            let seq_length = seq_lengths[current_batch];
+            let seq_length = seq_lengths[current_batch as usize];
             let total_seq_length = causal_mask.size()[1];
             let start_batch_size_times_num_heads = current_batch * config.n_head;
             let end_batch_size_times_num_heads = start_batch_size_times_num_heads + mini_batch_size * config.n_head;
